requirements.txt
altair==5.5.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
annotated-types==0.7.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
anyio==4.9.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
attrs==25.3.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
blinker==1.9.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
cachetools==5.5.2 ; python_full_version >= "3.9.8" and python_version < "3.13"
certifi==2025.1.31 ; python_full_version >= "3.9.8" and python_version < "3.13"
charset-normalizer==3.4.1 ; python_full_version >= "3.9.8" and python_version < "3.13"
click==8.1.8 ; python_full_version >= "3.9.8" and python_version < "3.13"
colorama==0.4.6 ; python_full_version >= "3.9.8" and python_version < "3.13" and platform_system == "Windows"
exceptiongroup==1.2.2 ; python_full_version >= "3.9.8" and python_version < "3.11"
fastapi==0.115.12 ; python_full_version >= "3.9.8" and python_version < "3.13"
gitdb==4.0.12 ; python_full_version >= "3.9.8" and python_version < "3.13"
gitpython==3.1.44 ; python_full_version >= "3.9.8" and python_version < "3.13"
h11==0.14.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
idna==3.10 ; python_full_version >= "3.9.8" and python_version < "3.13"
jinja2==3.1.6 ; python_full_version >= "3.9.8" and python_version < "3.13"
joblib==1.4.2 ; python_full_version >= "3.9.8" and python_version < "3.13"
jsonschema-specifications==2024.10.1 ; python_full_version >= "3.9.8" and python_version < "3.13"
jsonschema==4.23.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
markupsafe==3.0.2 ; python_full_version >= "3.9.8" and python_version < "3.13"
narwhals==1.32.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
numpy==1.26.4 ; python_full_version >= "3.9.8" and python_version < "3.11"
numpy==2.2.4 ; python_version >= "3.11" and python_version < "3.13"
packaging==24.2 ; python_full_version >= "3.9.8" and python_version < "3.13"
pandas==2.2.3 ; python_full_version >= "3.9.8" and python_version < "3.13"
pillow==11.1.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
protobuf==5.29.4 ; python_full_version >= "3.9.8" and python_version < "3.13"
pyarrow==19.0.1 ; python_full_version >= "3.9.8" and python_version < "3.13"
pydantic-core==2.33.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
pydantic==2.11.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
pydeck==0.9.1 ; python_full_version >= "3.9.8" and python_version < "3.13"
python-dateutil==2.9.0.post0 ; python_full_version >= "3.9.8" and python_version < "3.13"
python-dotenv==1.1.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
pytz==2025.2 ; python_full_version >= "3.9.8" and python_version < "3.13"
referencing==0.36.2 ; python_full_version >= "3.9.8" and python_version < "3.13"
requests==2.32.3 ; python_full_version >= "3.9.8" and python_version < "3.13"
rpds-py==0.24.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
scikit-learn==1.6.1 ; python_full_version >= "3.9.8" and python_version < "3.13"
scipy==1.13.1 ; python_full_version >= "3.9.8" and python_version < "3.11"
scipy==1.15.2 ; python_version >= "3.11" and python_version < "3.13"
six==1.17.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
smmap==5.0.2 ; python_full_version >= "3.9.8" and python_version < "3.13"
sniffio==1.3.1 ; python_full_version >= "3.9.8" and python_version < "3.13"
starlette==0.46.1 ; python_full_version >= "3.9.8" and python_version < "3.13"
streamlit==1.44.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
tenacity==9.0.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
threadpoolctl==3.6.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
toml==0.10.2 ; python_full_version >= "3.9.8" and python_version < "3.13"
tornado==6.4.2 ; python_full_version >= "3.9.8" and python_version < "3.13"
typing-extensions==4.13.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
typing-inspection==0.4.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
tzdata==2025.2 ; python_full_version >= "3.9.8" and python_version < "3.13"
urllib3==2.3.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
uvicorn==0.34.0 ; python_full_version >= "3.9.8" and python_version < "3.13"
watchdog==6.0.0 ; python_full_version >= "3.9.8" and python_version < "3.13" and platform_system != "Darwin"
